refactor(web-ui): deleted index barrel files from components, hooks and unit tests
