+++ packages/frontend/web-ui/src/app/dom/main.jsx
@@ -3,8 +3,8 @@ import ReactDOM from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
 import { Provider } from 'react-redux';
 import { store } from '../store/store';
-import { CornieAppTheme } from '../theme';
-import { CornieAppRoutes } from '../routes';
+import { CornieAppTheme } from '../theme/CornieAppTheme';
+import { CornieAppRoutes } from '../routes/CornieAppRoutes';
 
 import '../../scss/app.scss';
