--- packages/frontend/web-ui/src/app/routes/index.js
@@ -1,1 +1,0 @@
-export * from './CornieAppRoutes';
