--- packages/frontend/web-ui/src/app/theme/index.js
@@ -1,2 +1,0 @@
-export * from './CornieAppTheme';
-export * from './cornieTheme';
