--- packages/frontend/web-ui/src/auth/hooks/index.js
@@ -1,2 +1,0 @@
-export * from './useRegisterForm';
-export * from './useLoginForm';
