--- packages/frontend/web-ui/src/auth/pages/index.js
@@ -1,3 +1,0 @@
-export * from './Login';
-export * from './Register';
-export * from './RegisterConfirm';
