--- packages/frontend/web-ui/src/common/helpers/index.js
@@ -1,4 +1,0 @@
-export * from './validateFormConfirmPassword';
-export * from './validateFormEmail';
-export * from './validateFormName';
-export * from './validateFormPassword';
