--- packages/frontend/web-ui/src/common/hooks/index.js
@@ -1,1 +1,0 @@
-export * from './useShowPassword';
