--- packages/frontend/web-ui/src/common/http/helpers/index.js
@@ -1,1 +1,0 @@
-export * from './buildSerializableResponse';
